+++ src/quantity.rs
@@ -302,7 +302,7 @@ impl<U: Unit, V: Value> Quantity<U, V> {
 }
 
 impl<U: Unit, V: Value> Quantity<crate::units::UnitPow<U, typenum::P1>, V> where
-    U::Dim: crate::dimension::DimPowType<typenum::P1>,
+    U::Dim: crate::dimension::CanDimPowType<typenum::P1>,
 {
     /// Return an equivalent quantity with its exponent removed.
     pub fn cancel_exponent(self) -> Quantity<U, V> {